# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organisations/{organisationId}/integrations/providers": {
      "get": {
        "operationId": "Integrations_providers",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Providers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:read"
        ],
        "summary": "Registered integration providers",
        "description": "**Requires permission:** `integration:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/integrations/connections": {
      "get": {
        "operationId": "Integrations_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Connections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionList"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:read"
        ],
        "summary": "Connections (credentials are never returned)",
        "description": "**Requires permission:** `integration:read`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      },
      "post": {
        "operationId": "Integrations_create",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Connected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Connection"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:manage"
        ],
        "summary": "Connect a provider (credentials are validated then stored encrypted)",
        "description": "**Requires permission:** `integration:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConnectionRequest"
              }
            }
          }
        }
      }
    },
    "/api/v1/organisations/{organisationId}/integrations/connections/{connectionId}": {
      "delete": {
        "operationId": "Integrations_revoke",
        "parameters": [
          {
            "name": "connectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked"
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:manage"
        ],
        "summary": "Revoke a connection and wipe its credentials",
        "description": "**Requires permission:** `integration:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/integrations/connections/{connectionId}/check": {
      "post": {
        "operationId": "Integrations_check",
        "parameters": [
          {
            "name": "connectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:manage"
        ],
        "summary": "Health-check a connection",
        "description": "**Requires permission:** `integration:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ]
      }
    },
    "/api/v1/organisations/{organisationId}/integrations/connections/{connectionId}/execute": {
      "post": {
        "operationId": "Integrations_execute",
        "parameters": [
          {
            "name": "connectionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Makes retries safe: a repeated request with the same key and body replays the first response (24h).",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAccepted"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "401": {
            "$ref": "#/components/responses/401"
          },
          "403": {
            "$ref": "#/components/responses/403"
          },
          "404": {
            "$ref": "#/components/responses/404"
          },
          "422": {
            "$ref": "#/components/responses/422"
          }
        },
        "tags": [
          "Integrations"
        ],
        "x-required-permissions": [
          "integration:manage"
        ],
        "summary": "Queue a provider operation (never executed inside the request)",
        "description": "**Requires permission:** `integration:manage`",
        "security": [
          {
            "bearer": []
          },
          {
            "cookie": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExecuteIntegrationRequest"
              }
            }
          }
        }
      }
    },
...
  },
  "info": {
    "title": "UK Accounting Platform API",
    "description": "Practice-first accounting, tax and compliance platform. **V0 foundation:** identity, tenancy, RBAC, documents, jobs, audit, outbox events, workflows, tasks, notifications, integration and AI abstractions.\n\nAll tenant data lives under `/organisations/{organisationId}` and is protected by authentication, RBAC, company scope and PostgreSQL row-level security. Errors use RFC 9457 `application/problem+json`.",
    "version": "1.0.0",
    "contact": {},
    "x-api-versioning": "URI versioning (/api/v1). Additive changes only within a version; breaking changes ship as /api/v2."
  },
...
  "servers": [
    {
      "url": "/",
      "description": "Same origin"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Opaque session token from `POST /auth/login/bearer`"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "uk_session",
        "description": "httpOnly SameSite=Strict cookie set by `POST /auth/login`. Unsafe methods must send an allowed `Origin`."
      }
    },
    "schemas": {
      "Problem": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code"
        ],
        "properties": {
          "type": {
            "type": "string",
            "example": "urn:uk-platform:error:validation_error"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "code": {
            "type": "string",
            "description": "Stable machine-readable error code"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "instance": {
            "type": "string"
          },
          "correlationId": {
            "type": "string",
            "description": "Echoes `x-request-id`"
          }
        }
      },
...
      "JobAccepted": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "jobId",
          "status"
        ],
        "additionalProperties": false
      },
...
      "Connection": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "provider": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "REVOKED",
              "ERROR"
            ]
          },
          "companyId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "scopes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastCheckedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lastError": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "provider",
          "displayName",
          "status",
          "companyId",
          "scopes",
          "createdAt",
          "lastCheckedAt",
          "lastError"
        ],
        "additionalProperties": false
      },
...
      "CheckResult": {
        "type": "object",
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "detail": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          }
        },
        "required": [
          "ok",
          "provider"
        ],
        "additionalProperties": false
      },
...
      "ConnectionList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "provider": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "status": {
                  "type": "string",
                  "enum": [
                    "ACTIVE",
                    "REVOKED",
                    "ERROR"
                  ]
                },
                "companyId": {
                  "type": "string",
                  "format": "uuid",
                  "nullable": true
                },
                "scopes": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                "createdAt": {
                  "type": "string",
                  "format": "date-time"
                },
                "lastCheckedAt": {
                  "type": "string",
                  "format": "date-time",
                  "nullable": true
                },
                "lastError": {
                  "type": "string",
                  "nullable": true
                }
              },
              "required": [
                "id",
                "provider",
                "displayName",
                "status",
                "companyId",
                "scopes",
                "createdAt",
                "lastCheckedAt",
                "lastError"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
      "ProviderList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "provider": {
                  "type": "string"
                },
                "displayName": {
                  "type": "string"
                },
                "capabilities": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "provider",
                "displayName",
                "capabilities"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "items"
        ],
        "additionalProperties": false
      },
...
      "CreateConnectionRequest": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 60
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "companyId": {
            "type": "string",
            "format": "uuid"
          },
          "credentials": {
            "type": "object",
            "additionalProperties": {}
          }
        },
        "required": [
          "provider",
          "displayName",
          "credentials"
        ],
        "additionalProperties": false
      },
      "ExecuteIntegrationRequest": {
        "type": "object",
        "properties": {
          "operation": {
            "type": "string",
            "maxLength": 100
          },
          "params": {
            "type": "object",
            "additionalProperties": {},
            "default": {}
          }
        },
        "required": [
          "operation"
        ],
        "additionalProperties": false
      },
...
    },
    "responses": {
      "400": {
        "description": "Malformed request",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "401": {
        "description": "Not authenticated or session expired",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "403": {
        "description": "Authenticated but not permitted (RBAC, segregation of duties, CSRF origin)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "404": {
        "description": "Not found — also returned for resources in other organisations (existence is never revealed)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
      "422": {
        "description": "Validation failed (`errors[]` lists fields)",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearer": []
    },
    {
      "cookie": []
    }
  ]
}